Add files via upload
This file demonstrates how to extract Strings from a text file, one character at a time.  Each line of a text file gets stored in a String object, which in turn is stored in an ArrayList. For the EVG project, this will need to be adapted so that instead of being stored in an Arraylist, the Strings are used to construct Scenario objects.